+++ example/buttons.js
@@ -1,5 +1,12 @@
 var bot = require('../'),
-  buttons = [new bot.Button(bot.pins.p9_23), new bot.Button(bot.pins.p9_24)];
+  buttons = [
+    new bot.Button(bot.pins.p9_23),
+    new bot.Button(bot.pins.p9_24, {
+      debounceInterval: 0,
+      pullType: bot.pullTypes.PULL_DOWN,
+      isActiveLow: false
+    })
+  ];
 
 buttons.forEach(function (button) {
   button.on('error', function (err) {